+++ jkube-kit/generator/java-exec/src/test/java/org/eclipse/jkube/generator/javaexec/JavaExecGeneratorMainClassDeterminationTest.java
@@ -135,7 +135,7 @@ class JavaExecGeneratorMainClassDeterminationTest {
     void testMainClassDeterminationFromFatJar(@TempDir File baseDirectory) {
         try (MockedConstruction<FatJarDetector> fatJarDetector = mockConstruction(FatJarDetector.class,
                  (mock, ctx) -> {
-                     File fatJarArchive = new File("fat.jar");
+                     File fatJarArchive = new File(baseDirectory,"fat.jar");
                      when(fatJarDetectorResult.getArchiveFile()).thenReturn(fatJarArchive);
                      when(mock.scan()).thenReturn(fatJarDetectorResult);
                  });