+++ scripts/quickstarts.sh
@@ -24,7 +24,7 @@ function mvnVersion() {
   echo "Updating quickstarts pom files to JKube $JKUBE_VERSION"
   cd "$QUICKSTARTS" || exit 1
   find . -type f -name "pom.xml" -printf '%h\0' | xargs -0 -P 1 -I{} sh -c \
-    "cd {} && mvn -nsu -N versions:set -DnewVersion=$JKUBE_VERSION -DgenerateBackupPoms=false && cd ${QUICKSTARTS} || exit 255"
+    "echo Setting version for {} && cd {} && mvn -nsu -N versions:set -DnewVersion=$JKUBE_VERSION -DgenerateBackupPoms=false && cd ${QUICKSTARTS} || exit 255"
 }
 
 function gradleVersion() {
@@ -60,7 +60,8 @@ function packageMaven() {
 function packageGradle() {
   echo "Packaging all Gradle quickstart projects (excluding sub-modules)"
   cd "$QUICKSTARTS" || exit 1
-  find . -type f -name "build.gradle" -execdir ./gradlew build \;
+  find . -type f -name "gradlew" -printf '%h\0' | \
+    xargs -0 -L 1 -P 1 -I{} sh -c "echo Packaging {} && cd {} && ./gradlew clean build && cd ${QUICKSTARTS} || exit 255"
 }
 
 function package() {